feat: add models and update communications

+++ Assets/_Scripts/Communications/Mcp/AddMcpRequest.cs
@@ -4,8 +4,8 @@ namespace Communications.Mcp
 {
     public class AddMcpRequest
     {
-        [JsonProperty("Capacity")] public double Capacity { get; set; }
-        [JsonProperty("CurrentLoad")] public double CurrentLoad { get; set; }
+        [JsonProperty("Capacity")] public float Capacity { get; set; }
+        [JsonProperty("CurrentLoad")] public float CurrentLoad { get; set; }
         [JsonProperty("Latitude")] public double Latitude { get; set; }
         [JsonProperty("Longitude")] public double Longitude { get; set; }
     }
